phase-II/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
pyjwt==2.8.0
bcrypt==4.1.2
sqlmodel==0.0.22
psycopg2-binary==2.9.10
